reset the assigments as early as possible
Otherwise an old callback might be called when the user doesn't expect
it. The IO style makes it look like only one event at a time is being
waited upon.

Of course, waiting for only one event is a silly restriction, we'll just
have to come up with a different interface for doing it.