EventHook
Previously, all the event types were defined by their event-catching
procedure, which only catches a single event at a time. This will cause
problems later, when I we attempt to catch one of several types of
event.

With the new design, event types are described in terms of more general
events, and withNextEvent is defined in terms of event types instead of
the other way around.